Add very large grid database
This database is 20m x 20m. The problem with the smaller grid, as
pointed out by @tenxlenx, is that some of the routes go outside of it.

I went with a height of 1.8m (about 30cm off the ground), which is a bit
annoying, but with a lower height the database would have ended up
including images that were underground (the ground isn't level). We can
always replace this database with a better one once I figure out how to
use the ground as a mesh.